Coding Challenge Submission
I spent the full 4 hours on this submission. I originally wanted to create a loop that would inquire the user if they had further roman numerals to convert, however I was running out of time and decided to forego that idea in favor of completing the specified prompt. I also was not able to fully comment on my more complex functions. I have never used GitHub before, so I did need some time to learn how to use the website and what files were needed to be uploaded. I did not include that time into the four hours.